RE #27346 Make scripts run without errors
Scripts containing an encoding declaration in the first two lines are
now able to be run without errors or encoding artefacts.

As the declaration cannot compiled properly is not encoded and causes
encoding artefacts if other parts containing unicode characters are
encoded, the solution has been to remove the declaration from the code
to be run if there would be a failure.