Move matplotlib methods CurveTabWidgetPresenter to FigureErrorsManager
This commit moves all functions in the CurveTabWidgetPresenter that only use
matplotlib functions into the figureerrormanager. This is done because the
CurveTabWidgetPresenter is a presenter for the settings tab on the Mantid plot
window and is not suitable to manage plotting outside its scope, while
figureerrormanager already handles plotting functions for the MantidPlot window.

Re #26330