Move tests for figureerrorsmanager to test_figureinteraction
This commit moves tests in figureerrorsmanager that relate to the functions moved
in commit 932fc9. These tests relate to the generation of the error bar context menu.

This commit also changes the context menu text in figureinteraction from global variables into class attributes, as well as the relevant changes to any references to these attributes.

Re #26330