Move tests for curvestabwidgetpresenter to test_figureerrorsmanager
This commit moves tests in curvestabwidgetpresenter that relate to the functions
moved in commit 86c9e1. These tests relate to the the replotting of multiple curves.

This commit also removes the context menu text attribute from FigureErrorsManager that are no longer used since the context menu generator functions were moved to the FigureInteraction class in commit 932fc9

Re #26330