Abins broadening: refactor interpolation, apply PEP8
Fast interpolated broadening is now moved into a more
general/extensible function, including ascii-art documentation.
sqrt(2) spacing is implemented, and initial testing shows that this
noticably improves accuracy for little extra cost. Set this as the
default, and for now leave in the coarser option for test purposes.

A flag is also provided to skip rebinning inside this function; this
potentially avoids and expensive redundant operation but makes for
convoluted logic when this rebin is also considered inside the
Instrument. A cleaner scheme should be possible, in which it is
obvious where rebinning should happen.