Accept OrientedLattice by unique_ptr
The overwhelming use case for this method is to
take ownership of a lattice object into Sample.
Using unique_ptr clarifies that this is happening
at the call site and forces a caller to make copy
if that is truly the desired behaviour. The old code
make a copy internally but this was not obvious
from the method call.