Updated handling of resolution in convFit Re #26881
This updates the handling of which resolution spectra to use for each spectra in the fit. This needed updating because for multiple files it was using the convolution for the first file for all files and because how QENSFitSequential and QENSFitSimultaneous handle convolutions is different. The sequential version requires a single workspace which contains all of the resolutions whereas the simultaneous version requires the resolutions to be encoded into the FitFunction.