Re #25358 Fix edge cases with serializer
Fixes up several edge cases that the serializer was running into. This
includes: a list of enums not being serialized, and a class<->instance
mismatch.

The latter manifests itself subtly where setting the class attr then
changing the instance attr will actually result in different things.
This caused the LARMOR system test to be out by a magnitude of scale as
the wrong shape was used.

Additionally, we now throw a runtime error when we get something that
isn't serializable to help continue to flush out any other edge cases
that are left