Bump Minimum CMake to 3.13
Bumps the minimum CMake version to 3.13, this allows us to use
add_link_options on address sanitizers.
At the time of this commit 3.14 is available for RHEL7, and 3.15 for
Ubuntu. Windows requires 3.15 so in an ideal world we would just set our
minimum to that, but as it's not fully available we settle for the
earlier version at the moment.