Add workflow algorithm for merging workspaces with limits
This commit creates a workflow algorithm named MergeWorkspacesWithLimits
that handles the merging of the focused workspaces in the polaris scripts.

This commit removes a duplicated _load_qlims function in polaris_algs mistakenly added in commit 4c588e.

This commit adds a black line at the end of the CalculateSelfScatteringCorrection algorithm.

Re #27445