Add Document page for MergeWorkspacesWithLimits
This commit creates a documentation rst file for MergeWorkspacesWithLimits
Along with the workflow diagram dot file. It also makes changes to the
MergeWorkspacesWithLimits algorithm to include a name and version that
is required by the documentation compiler.

This commit corrects a mistake in the property names for
CalcuateSelfScatteringCorrection.

This commit removes the commented out command to the
_merge_workspaces_with_limits function in polaris_algs in place of the workflow
algorithm.

This commit creates a blank rst file that will later contain the documentation for
CalculateSelfScatteringCorrection.

Re #27445