Write-up Abins interpolated broadening scheme for developer docs
Figures are included as image files and the code to generate them is
included as a comment. This doesn't seem ideal as they may be revised
and bloat the repo (or become outdated); is there any workflow to
generate images on the fly when building the docs?

The section on interpolation method is a bit long and should perhaps
be spun out into another document.