Add document for CalculateSelfScatteringCorrection
This commit creates a documentation rst file for CalculateSelfScatteringCorrection
Along with the workflow diagram dot file. It also makes changes to the
CalculateSelfScatteringCorrection algorithm to include a name and version that
is required by the documentation compiler.

this commit add periods to the end of each set in the sub algorithm list in
MergeWorkspaceWithLimits.

this commit corrects copyright dates in the CalculateSelfScatteringCorrection and
MergeWorkspacesWithLimits algorithms.

Re #27445