Add option to specify which filter algorithm to use
There are differences between the old FilterByTime algorithm and the new
FilterEvents algorithm. This commit adds an option so the user can
specify which to use and sets the default to use the old one.

Also, when combining the monitors and detectors into a single workspace, we
were merging the logs, which meant the proton charge is different to
previous behaviour when scientists were doing this step manually. This
commit changes MergeLogs to be false in line with the way scientists
previously did this.

I'm not sure this is really an issue because the final proton charge for
the workspace is taken from the monitor workspace, which contains the
unfiltered proton charge. We may want to adjust this in future to be the
proton charge for the filtered slice.

Re #25881