Add further notes to changes in release notes
This commit adds to the lease notes that the fix for the bug that raised an error when a unicode string was given as a lim file directory has been fixed.

This commit adds to the lease notes that the match spectra in merging workspaces no longer matches by scale.

Re #27444