Abins broadening: rework normalisation to work with extreme values
Normalise Gaussians by multiplying by bin-width instead of dividing by
sum(values). The sum approach can go badly wrong if values lie near or
beyond the sampling range, as as tail fragment (or noise floor!) would
be amplified to unity. Scaling by the theoretical value as we do here
has a different downside: truncated broadening kernels will not quite
sum up to 1, so a little intensity is lost in broadening. This error
is more tolerable, especially as it can be decreased by extending the
truncation range.