Adding tiled plotting by pairs and groups
Based on the feedback from a scientist the home tab group widget is being removed.
Plotting is now controlled in the grouping tab

The tables in the grouping table now have analyse column, which decides whether the group or pair will be plotted.
Adding option 'tiled by'

This enables the plots to be tiled by run or group/pair

Plot will now auto update its layout on resize

Fitting tab is updated to abide by the groups and pairs selected in the grouping tab

Tiled plotting scales to the same axis

Separating dock widget into its own file such that the window layout will update when the window is docked and undocked