RE #27582 Make config threads compiler dependent
Windows devices have been running into issues in places where calling
the functions created to fix an issue on Unix machines in parallel
regions was causing error messages to  appear.

This is due to changes in implementation of omp_set_num_threads between
OpenMP v2 and v2.5+. As MSVC uses the older version of OpenMP I have
added a check that does not bother with the changes from the previous
fix if compiled with MSVC, as the original issue only appeared on non
MSVC compiled versions anyway.