PlotPeakByLogValue takes a multidomain function input Re #26881
PlotPeakbyLogValue performs a batch fit across the spectra of a workspace. It uses the same fitting function for each of these but with the option of updating any Attribute called WorkspaceIndex to match the workspace index of the input spectrum. This is used for example to use the same fitting function but with different resolutions for each

It also possesses the ability to give different values to the Q attribute once again referenced by workspace index but this time controlled by the special Qdependent function subtype. Neither of these methods really work when using the multi inputs options in the IndirectDataAnalysis tab as then you have input spectra from multiple files and the reliance on the input workspace index stops working.

To work in this more complex situation PlotPeakByLogValue has been updated to optionally take a multidomain function and then split it up which allows more fine grained control of what is being fitted.