RE #27187 Add full calibration functionality
Currently, there is no way to load a CSV file (which is the format
that the full calibration is in) into mantid to be used by the GUI.
Until that is added, the option to select the full calibration has been
hidden from the user.
The functionality is there, so once the loading works for large CSVs the
hide() call can be removed.