Update slicing test reference file
The new reference file has the following changes:
- The IvsQ_binned results contain slightly different values due to
binning differences to the original tests which did a manual Rebin of
the detectors and monitors workspaces. The new workflow algorithm just
rebins the detectors the same as the monitors. If I hack the script to
use the old rebin params I get exactly the same results, so this
confirms that functionally the new script is correct.

- The new reference includes some interim workspaces (e.g. IvsQ and
TOF) which were deleted in the original test scripts. These are standard
outputs so I think it's better to include them.

- The test script has been updated to not delete the interim workspaces
mentioned above. It also now does delete the interim TRANS_LAM
workspaces, because the workflow algorithms currently always output
these for input workspace groups, even though they shouldn't really if
debug is not on.

Re #25881