Refactor the inner loop in splitByTimeVector
The inner loop had a section to check the next value in the loop, which
did the same as the next iteration of the main loop would do anyway.
This commit consolidates it to just proceed to the next iteration in the
loop.

Also, the outer loop was originally split into two separate loops due to
needing to continue if the inner loop ran out of values first. This
commit also refactors the inner loop such that the outer loop can be
performed in a single loop.