Output TOF workspaces individually rather than by group
Outputting the group is fine if the group does not already exist, but if
the group exists then I had trouble outputting only the new members of
the group, and it's inconsistent which outputs you get depending on the
state before the algorithm was run. I therefore think it's simpler to
output workspaces when they are created. If the group does not already
exist in the ADS then it needs to be added and I've had to use the ADS
directly for this because we can't set the group as an output property
if its members are already output or their history will be duplicated.

Re #27633