Disabled undefined-var-template warning Re #26881
This warning is complaining that the static type g_typeMap does not have
explicit instantiations in this translation unit. We have explcict
instantions of all the template specialisations we require in the cpp file
and if someone adds one later it will lead to a linker error so think we are
ok to disable this warning.