Remove deprecated toolbar attributes
As far as I can tell these were leftover attributes from
a very old version of the matplotlib NavigationToolbarQT
class that have now been deprecated in matplotlib 3.1 and
set as readonly. A search back through matplotlib 1.5
suggests that they were not used and 1.5 is the minimum we
are required to support.