Update Abins tests, fix CASTEP tests
Test data needs updating due to changes in the Advanced
Parameters. Have eyeballed some of the file changes and the changes
seem fine. (There's also an increase in the written precision of atom
positions, but that will be related to the new test data writer.)

CASTEP data updated here, other codes to follow.