Abins: Disable quantum orders 3,4
The accuracy of results obtained with QuantumOrderEventsNumber > 2 is
suspect; in hydrogenous materials the magnitudes of order 3 and 4
appear surprisingly large, and it is not clear that the removed
thresholding procedure would give results that are correct.

While the origin of this behaviour is investigated, these orders are
disabled in the Abins algorithm as they cannot be recommended for use
in production calculations.

The underlying code is left in place, as it is likely that a corrected
procedure will

- closely resemble the existing one
- be implemented in the near future

If that work is abandoned then the code for higher orders should be removed.