Allow ProcessingAlgorithm to use old version of algorithms
Changes related to this PR have caused test failures in the
data processor UI because they use one of the changed
reflectometry algorithms. Reflectometry no longer uses the
data process, so this is somewhat obsolete. In order to maintain
working tests I've added an option to specify the version of the
algorithm to create in the ProcessingAlgorithm so that we can maintain
the previous behaviour and previous tests.

Re #27633