Add new version of the base class to preserve old properties
This commit adds a new version of ReflectometryWorkflowBase, which is
the base class for the reflectometry reduction algorithms and does the
work to set up the transmission properties. The old version is required
to preserve the original properties for version 2 of
ReflectometryReductionOneAuto. The new version inherits from version 2
and just overrides the function to set up the transmission properties.

Re #27633