RE #27659 Make EnggCalibrateFull use SaveAscii
Output format is a TSV file, due to the commas present in the lists.

Output for the detector positions table is now saved via the SaveAscii
algorithm so that the file can be loaded back into mantid and used in
calibrations and focusing. The original used an unloadable output
format.