Simplify output properties and grouping
The new code to declare properties on the fly was causing too many
problems so I've removed it. I've added proper output properties for the
transmission outputs from the child algorithms. The other output
properties I've removed. It means these workspaces will not have history
but this is low priority because they are the input TOF workspaces so
the history is not critical. It would be nice to have it for the
summed/sliced workspaces, but this can always be extracted from the main
outputs.

I've also removed the grouping of additional TOF workspaces, so
behaviour is now reverted to the original where only the loaded TOF
workspaces are grouped.

Re #27633