Changed test for signal to noise ratio with background
Altered test to generate random flat background in spherical region and testing fractional difference against an expected calculated value (rather than hard-coded values). This did mean adding an optional argument to estimateSignalToNoiseRatio in Integrate3DEvents.cpp to force the region to be spherical if the fractional difference of min and max sigma is within some tolerance (needed to compare to expected value). These optional parameters are not used anywhere else in the code.