Re #27795 Don't act differently on 0 offset m5
In a previous commit I added we treated 0. offset as a special case of
doing nothing, as naively I assumed the IDF position of M5 is correct

However, this skips the logic which repositions the monitor at the rear
detector, then applies a shift. So a shift in rear-detector relative to
the monitor IDF position would result in a different result.

This lead to confusing behaviour and incorrect results when a shift of 0
is set.