Fix problems with output property lists
When assigning to a list of outputs when calling the algorithm in
python, the outputs must be specified in the order they are declared,
even if they are optional outputs like IvsLam. This means that if we
want to specify the transmission workspace output we must first specify
IvsLam even if debug is false. Some of the algorithms always output
IvsLam if the name was given whereas others did not. They now always
output it if the name is given.

Note that this means you may now get an error if you don't supply enough
arguments whereas previously it was ok. It is ok if you just specify the
IvsQ output(s) but beyond that we do not know whether the output should
be IvsLam or the transmission, so we require all of them, hence the
error. This is better than it assigning an unexpected output so I think
is ok, and is the only way I can think to do this.

Re #27633