Add switch to allow the same tracks to be reused for different lambas
A switch has been added to the MCAbsorptionStrategy class that allows a pair
of before scatter\after scatter tracks to be used to calculate the attenuation
for a range of different lambda values. This will improve the speed of the Monte
Carlo calculation. The Monte Carlo calculation currently sets the switch to
reuse the tracks without exposing this in the algorithm parameters - I've mainly
left the "old" approach in place in case any further wavelength dependent effects
are added to the absorption calculation eg beam spreading

In order to add the reuse while retaining the original behaviour (via the switch),
I had to reverse the order of the event and lambda loops. The previous code looped
over lambda with an inner loop of events. The event loop is now the outer loop, so
even with the switch set to not reuse tracks the results will change slightly if
the simulation isn't converged

The function MCInteractionVolume::calculateAbsorption has been split into two:
- track generated done in calculateBeforeAfterTrack
- absorption calculation done in calculateAbsorption

The relevant unit and system tests have been adjusted. In particular the tests in
MCInteractionVolumeTest have been split out so the track generation and absorption
calculation are tested separately