Re #27028 Fix tabbing issues in SANS GUI
Fixes tabbing issues in SANS GUI, this requires major structural changes
to the types passed around the GUI. Previously, the codebase used an
array of strings, and relied on the position

We have swapped out this with a RowEntries object which attempts to be
as close to a POD type as possible. With additional refactoring later it
should be complete.

As the type is very pervasive replacement touches a significant portion
of the code base.