Re #27028 Fix BatchProcessing using state IDS for table model
Previously each state object could be back linked to the table model, by
tying an ID to the row. This is brittle as it links the indexing method
of the table to the rest of the program.

Additionally, we now produce a tuple of row, index pair to help go
through the Qt signal mechanism