Use algorithm documentation for tooltips on the table
This commit improves the tooltips on the per-theta settings table by
setting a separate tooltip for each column. Previously, all columns had
the same tooltip. The tooltips are taken from the algorithm
documentation for the relevant property.

Note that the tooltips are only set on the table cells. The table
headers all have the original default tooltip, which explains the
context of what the table does.

This is a useful change in itself but is also preparative work for issue
27826 which will add further complexity to the per-theta defaults table,
making the tooltips more important.

Note that the underlying algorithm that is being used has been changed
to ReflectometryISISLoadAndProcess. This change should have been made
earlier when that algorithm was added but this was not spotted because
the new properties in that algorithm were not needed before now.

Re #27826