Addition of a sequential fitting tab
This tab will allow for sequential fits to be performed easily. This commit creates the tab, and wires it up such that the displayed workspaces are updated, and the fit function corresponds to that shown within the conventional fitting tab. Additionally, this commit wires the tab up for single fitting