Clean function for performing a fourier filter to PDF
This commit cleans and fixes many bugs in the fouriet filter function in polaris_algs as well as making the filter params passable into the create_total_scattering_pdf.
It also cleans up the polaris paramiter map so that some of the inputs are optional so the defaults can be used.

re: 27773