Re #27028 Update Batch Parser to handle new type
Updates the Batch CSV processor to handle the new type, this includes
switching the implementation to use enums. This additionally should fix
some bugs around the order of tabs being fixed in a CSV file, which
caused the sample shape toggle to break files if it didn't match the
saved file.