Update system tests with transmission binning fix
This commit updates the reference files for tests where the transmission
workspace is output from ReflectometryReductionOne. The refactored code
in this PR has introduced a fix in this algorithm to output the
transmission workspace after it has been rebinned to detectors.
Previously it was output before the rebin step, hence the minor
differences in the output. Note however that the final results of the
reduction (IvsQ workspaces) are unchanged.

This commit also moves the code to set the output workspace to the
bottom of the function that does the transmission correction, to help
avoid a similar problem happening again.

Re #27633