Re #27028 Fix QThreads in SANS Unit tests
Fixes QThreads getting mixed in with SANS Unit tests. run_tab_presenter
is the main offender as it reaches across into other presenters. For the
moment we mock out everything we can and try to do our best to ensure
that we aren't using QThreads

This means a race condition still exists where if the UI closes before
the thread pool is finished we will have the GIL throw. However, users
will not open, load and close the GUI < 1 second.

In the next release we will look at a project wide implementation in
Python 3 for handling thread pools, rather than various technique
specific ones