Re #27028 Update LOQ beam centre rmin max values
LOQ used to previously use rmin and rmax values from an IPF file, which
we had to determine on the run. These files (in the IPF) were wrong, and
the scientists were unaware of the existance of the IPF and basically
changed them anyway.

On discussion we can drop the IPF file loading (massively simplifying
the code), and hardcode these values for LOQ. In a future commit we will
hard-code optimal defaults for each instrument and bank position