RE #27779 Implement observer between tab widgets
Observer pattern uses has the data widget as an observable that sends
the plotting widget the workspaces to add or remove from the plot when
the 3rd column in the table is ticked and unticked respectively. The
pattern also handles the removal of indiviual workspaces or the
clearing of all tracked workspaces.