Create SaveRMCProfile
This commit creates an algorithm `SaveRMCProfile` to output an ascii file with the format corresponding to a file used by the RMCProfile package. the output file contains only the x and y data from a workspace with a header containing the number of y values in the file on the first line, and on the second line the type of function contained within and a title for the data.

re: #27774