Re #27663 Add new commands to ISISCommandInterface
Adds two new commands to the ISISCommandInterface (legacy) :
- SetMonDirect which sets the correction file
- SetDetectorMaskFile which sets the masking file (not detector masks)

These are not added to the release notes as the ISISCommandInterface is
deprecated. However, these calls are littered around some tube calib
scripts which are being prototyped.

On weighing up the balance, we will add them to make it easier to later
merge the calib scripts. Since similar functionality is difficult to get
at using the new command interface.