Updated test values in system test
The stdDev function that was deleted in this branch often returns different values  than the 'correct' value which is sqrt(eigenvalue)  - which is now used. Typically the difference is  <0.5%, however occasionally the stdDev function returned a std deviation that is unrealistically small (by two orders of magnitude), this behavior is not as yet understood, but since the function is no longer used it is appropriate to change the test values.